Move the location of addLegacyBookmarkToUIFolder to the test resources folder
